move Trekpedia class to separate file
